feat(ui/budget): move sub pages into own sub modules
