+++ ui/src/pages/budget/project.rs
@@ -1,0 +1,13 @@
+use avina_wire::user::UserDetailed;
+use dioxus::prelude::*;
+
+#[component]
+pub fn BudgetProjectSubPage(
+    api_url: String,
+    token: String,
+    user: UserDetailed,
+) -> Element {
+    rsx! {
+        p { "TODO" }
+    }
+}
